Add GitHub Actions workflow for running tests (#3)
Runs the tests and checks the coverage using a custom step instead of
Codecov. Based on https://hynek.me/articles/ditch-codecov-python/. Add a
dummy test so that `make test` doesn't fail because no tests were found.
Install the package in editable mode instead of building a wheel because
coverage can't find the source files otherwise. Coverage must be run
with relative paths in order to combine the results from different OSs
on CI.