Add a function to generate n-dimensional grids
This is a port of the `grid_coordinates` function in Verde. Here, the
`extra_coords` argument was renamed to `non_dimensional_coords` and the
`meshgrid` option was removed (it could be replaced by multiple calls to
`line_coordinates` if desired). The function was also generalized to
n-dimensional grids (depending on the number of dimensions in the region).