Change both start and stop when adjust region (#21)
Before, when asking `line_coordinates` to adjust the region instead of
spacing, it would only add the required padding to the upper boundary of
the region. It would make much more sense to pad the region on both
sides instead. Do this by calculating the padding required and adding it
to the interval given. Tests and documentation are updated accordingly.
This is a departure from the values originally given in the Verde
function.


**Relevant issues/PRs:** Fixes #17