Add a function to generate n-dimensional grids (#18)
This is a port of the `grid_coordinates` function in Verde. Here, the
`extra_coords` argument was renamed to `non_dimensional_coords` and the
`meshgrid` option was removed (it could be replaced by multiple calls to
`line_coordinates` if desired, so here we always meshgrid). The function
was also generalized to n-dimensional grids (depending on the number of
dimensions in the region). This meant switching the meshgrid to use
`indexing="ij"` because the `"xy"` indexing behaves strangely for more
than 2 dimensions. Also make the arguments keyword-only to avoid
confusion of passing `spacing` as the second positional argument and
getting wrong results.