Add function to check coordinates have same shape (#29)
The `check_coordinates` function was ported from Verde and made part of
the public API of Bordado. It also makes sure the coordinates are numpy
arrays so that we can use lists when passing coordinates to other
functions. Use the function in other functions and improve their
docstrings description of coordinates.