Add function to generate coords between 2 points
The `profile_coordinates` function was ported from verde and expanded.
Parameter values are more sensible. It now accepts and size or a spacing
and uses `line_coordinates` instead of `linspace`. Uses a directional
vector for the maths instead of calculating an angle so it can
generalize to more than 2 dimensions.