Remove validation functions from the public API
The `check_*` functions are used widely internally and we have put them
on the API in case others want to use them. But their specs can change
with time and maintaining compatibility would be a pain. Different
projects can also have different requirements of the inputs so these are
hard to generalize. Keep them private instead to avoid issues in the
future.