Add function to make points along a great circle
The `profile_coordinates` function is Cartesian but when dealing with
geographic data we may want to lay points along a great circle path on
the sphere. This function implements this functionality and works only
for (longitude, latitude) points. The radius of the sphere can be
customized but defaults to the mean Earth radius.