Add a function to generate random points on a sphere (#59)
Generates uniform randomly distributed points on the surface of a
sphere. Using the regular `random_coordinates` for this would lead to
denser sampling at higher latitudes. To achieve this, we sample the
cosine of latitude (in this case, colatitude) instead of the angles
themselves. Test the code by calculating the point density per latitude
band and making sure it's roughly uniform.